requirements.txt
click==6.7
Flask==1.0.2
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
mysql-connector==2.1.6
Flask-Session
